Separate and update maintained crates list (#8)
As mentioned some time ago during my meetup talk we should remove `zmq2`
from this list now that we've been able to publish our updates under the
original `zmq` name.  For the occasion, create a separate list with
crates that we actively maintain and add `ash` and the `rust-mobile`
crates, and extend the list of our crates with `vdb-rs`.